# web/openapi.json
{
  "components": {
    "schemas": {
      "DailyResponse": {
        "description": "Daily time-series response contract.",
        "properties": {
          "daily": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/DailyUsage"
            },
            "title": "Daily",
            "type": "array"
          },
          "pricing_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pricing Source"
          },
          "total_cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Cost Usd"
          },
          "total_sessions": {
            "default": 0,
            "minimum": 0.0,
            "title": "Total Sessions",
            "type": "integer"
          },
          "totals": {
            "$ref": "#/components/schemas/TokenUsage"
          },
          "window_days": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window Days"
          }
        },
        "required": [
          "totals"
        ],
        "title": "DailyResponse",
        "type": "object"
      },
...
      "DoctorReport": {
        "description": "Top-level health diagnostics payload.",
        "properties": {
          "app_name": {
            "title": "App Name",
            "type": "string"
          },
          "app_version": {
            "title": "App Version",
            "type": "string"
          },
          "legacy": {
            "$ref": "#/components/schemas/LegacyDiagnostics"
          },
          "opencode_data_dir": {
            "title": "Opencode Data Dir",
            "type": "string"
          },
          "selected_source": {
            "title": "Selected Source",
            "type": "string"
          },
          "sqlite": {
            "$ref": "#/components/schemas/SQLiteDiagnostics"
          }
        },
        "required": [
          "app_name",
          "app_version",
          "opencode_data_dir",
          "selected_source",
          "sqlite",
          "legacy"
        ],
        "title": "DoctorReport",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelDetailResponse": {
        "description": "Single model detail response contract.",
        "properties": {
          "cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd"
          },
          "daily": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/DailyUsage"
            },
            "title": "Daily",
            "type": "array"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "pricing_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pricing Source"
          },
          "total_interactions": {
            "default": 0,
            "minimum": 0.0,
            "title": "Total Interactions",
            "type": "integer"
          },
          "total_sessions": {
            "default": 0,
            "minimum": 0.0,
            "title": "Total Sessions",
            "type": "integer"
          },
          "usage": {
            "$ref": "#/components/schemas/TokenUsage"
          },
          "window_days": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window Days"
          }
        },
        "required": [
          "model_id",
          "usage"
        ],
        "title": "ModelDetailResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ModelMeter",
    "version": "2026.3.17"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/daily": {
      "get": {
        "operationId": "daily_api_daily_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "default": 7,
              "title": "Days"
            }
          },
          {
            "in": "query",
            "name": "timezone_offset_minutes",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 840,
              "minimum": -840,
              "title": "Timezone Offset Minutes",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "token_source",
            "required": false,
            "schema": {
              "default": "auto",
              "enum": [
                "auto",
                "message",
                "steps"
              ],
              "title": "Token Source",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "session_source",
            "required": false,
            "schema": {
              "default": "auto",
              "enum": [
                "auto",
                "activity",
                "session"
              ],
              "title": "Session Source",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "db_path",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Db Path"
            }
          },
          {
            "in": "query",
            "name": "pricing_file",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pricing File"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Daily"
      }
    },
    "/api/doctor": {
      "get": {
        "operationId": "doctor_api_doctor_get",
        "parameters": [
          {
            "in": "query",
            "name": "db_path",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Db Path"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Doctor"
      }
    },
...
    "/api/models": {
      "get": {
        "operationId": "models_api_models_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "default": 7,
              "title": "Days"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "db_path",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Db Path"
            }
          },
          {
            "in": "query",
            "name": "pricing_file",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pricing File"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Models"
      }
    },
    "/api/models/{model_id}": {
      "get": {
        "operationId": "model_detail_api_models__model_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "default": 7,
              "title": "Days"
            }
          },
          {
            "in": "query",
            "name": "db_path",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Db Path"
            }
          },
          {
            "in": "query",
            "name": "pricing_file",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pricing File"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Model Detail"
      }
    },
    "/api/projects": {
      "get": {
        "operationId": "projects_api_projects_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "default": 7,
              "title": "Days"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "db_path",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Db Path"
            }
          },
          {
            "in": "query",
            "name": "pricing_file",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pricing File"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Projects"
      }
    },
    "/api/projects/{project_id}": {
      "get": {
        "operationId": "project_detail_api_projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Days"
            }
          },
          {
            "in": "query",
            "name": "session_offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Session Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "session_limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Limit"
            }
          },
          {
            "in": "query",
            "name": "db_path",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Db Path"
            }
          },
          {
            "in": "query",
            "name": "pricing_file",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pricing File"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Project Detail"
      }
    },
...
    "/api/summary": {
      "get": {
        "operationId": "summary_api_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "default": 7,
              "title": "Days"
            }
          },
          {
            "in": "query",
            "name": "token_source",
            "required": false,
            "schema": {
              "default": "auto",
              "enum": [
                "auto",
                "message",
                "steps"
              ],
              "title": "Token Source",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "session_source",
            "required": false,
            "schema": {
              "default": "auto",
              "enum": [
                "auto",
                "activity",
                "session"
              ],
              "title": "Session Source",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "db_path",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Db Path"
            }
          },
          {
            "in": "query",
            "name": "pricing_file",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pricing File"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Summary"
      }
    },
...
  }
}